test/dataflow/source/requirements.txt
azure-functions
adlfs==2022.11.2
azure-core==1.26.1
azure-datalake-store==0.0.52
azure-identity==1.12.0
azure-storage-blob==12.14.1
cffi==1.15.0